Restructure image tests
