+++ tests/image_tests.rs
@@ -1,6 +1,7 @@
 use {super::*, std::sync::Once};
 
-mod browser_tests;
+mod browser;
+mod native;
 
 macro_rules! image_test {
   (
@@ -12,13 +13,13 @@ macro_rules! image_test {
 
       #[test]
       fn native() -> Result {
-        image_test(stringify!($name), $program)
+        native::test(stringify!($name), $program)
       }
 
       #[tokio::test]
       #[ignore]
       async fn browser() -> Result {
-        browser_tests::browser_test(stringify!($name), $program).await
+        browser::test(stringify!($name), $program).await
       }
     }
   };
@@ -32,12 +33,12 @@ macro_rules! image_test {
 
       #[test]
       fn native() -> Result {
-        image_test(stringify!($name), $program)
+        native::test(stringify!($name), $program)
       }
 
       #[tokio::test]
       async fn browser() -> Result {
-        browser_tests::browser_test(stringify!($name), $program).await
+        browser::test(stringify!($name), $program).await
       }
     }
   };
@@ -59,44 +60,6 @@ fn clean() {
       }
     }
   });
-}
-
-fn image_test(name: &str, program: &str) -> Result {
-  let destination = format!("images/{}.native-actual-memory.png", name);
-
-  fs::remove_file(&destination).ok();
-
-  let tempdir = Test::new()?.program(program).run_and_return_tempdir()?;
-
-  let actual_path = tempdir.path().join("memory.png");
-
-  let actual_image = image::open(&actual_path)?;
-
-  let expected_path = format!("images/{}.png", name);
-
-  if !Path::new(&expected_path).is_file() || actual_image != image::open(&expected_path)? {
-    fs::rename(&actual_path, &destination)?;
-
-    #[cfg(target_os = "macos")]
-    {
-      let status = Command::new("xattr")
-        .args(["-wx", "com.apple.FinderInfo"])
-        .arg("0000000000000000000C00000000000000000000000000000000000000000000")
-        .arg(&destination)
-        .status()?;
-
-      if !status.success() {
-        panic!("xattr failed: {}", status);
-      }
-    }
-
-    panic!(
-      "Image test failed:\nExpected: {}\nActual:   {}",
-      expected_path, destination,
-    );
-  }
-
-  Ok(())
 }
 
 image_test! {
